+++ lib/zenlish/wclasses/demonstrative_pronoun.rb
@@ -1,0 +1,10 @@
+require_relative 'pronoun'
+
+module Zenlish
+  module WClasses
+    # Demonstrative pronouns identify a noun: this, that, these, those, such
+    # or someone/something mentioned explicitly this one
+    class DemonstrativePronoun < Pronoun
+    end # class
+  end # module
+end # module